Fix errors in Temporal types serialization
Date types are not being decoded propely and failing with an error.
Time types were parsing wrongly milliseconds.
Duration was not handling formats with time.

This errors were find with the minimum functionality integration tests.
Theses tests are for testing basic functions in the driver.
Problem with the spatial types were also found, but their fix as defered
to another pull request.

Problem with the Map type was also found, but the driver behaves the
same.
So, investigation will be needed.

Graph types were not tested yet, but they will in a new PR.

The ESCMAScript version was bumped to 2020 since this version is needed
for supporting create bigint as literal.
This doesn't affect the minimal requirements of using the library since
Node 18 supports ES2020.