Remove support for DateTime without offset
The usage of this contruction is prune to errors like ambiguous times.
For instance, `2022-10-30T2:30:00[Europe/Berlin] could be GMT+1 or
GMT+2.`.
And unexisting times like `2024-03-31T02:30:00[Europe/Stockholm`.

The solution is not allowing the user send this kind of DateTime to the
server by thrown an exception when serialize it.
This approach is needed because we can't thrown on the DateTime creation
since it will configure a breaking change on the driver.
This library works slighty different is not an issue since the user need
to opt-in to the usage of this new library.