Introduce Resource Management
The wrapper should not create infinity number requests to the Database.
This number should be controlled using the `maxConnectionPoolSize`
configuration.
This should have similar behaviour as the connection pool in the
`Driver`, since both aims to limit the amount of connections and to
avoid to submit to much work to the database.

Along with the sizing limit, `connectionAcquisitionTimeout`
configuration is introduced to limit the amount of time a transaction
will wait for acquiring a connection.

It is important to noticed that connections object are just logic
wrappers for executing requests to the Database.
So, these objects will be created as needed, but they will last until
the wrapper get closed.
This avoids creating and destroying objects all the time, but it might
cost a bit more of memory usage.