Check auth on `Wrapper.verifyConnectivity`
The driver version of the method fails on invalid creadentials, so this
also have to do it for better interchangeable usage.

In the case of the Wrapper, a query might be run for verifying if the
database can handle requests from the driver credentials.

Along with this changes, other changes related to the endpoint
configuration were also done so the rediscovery get triggered when
endpoint is not accessible.