Adjusting support for Duration (#35)
This add support to `hour` and `minute` which are converted to seconds.
Support to `week` is also added, this is converted to days.

Integration tests using example code from Cypher Manual were added for making sure the wrapper is returning the same value as the driver in the example cases. 

Along with those changes, error messages got improved to get more hint about the limitations or errors.