Add support to explicit transactions
The explicit transaction enables the apis `Driver.executeQuery`,
`Session.beginTransaction`, `Session.executeRead` and
`Session.executeWrite`.
The access to the `Transaction` object is also granted.
All limitations around transactions in the drivers are also applicable
here.

In manner to transactions work in a cluster enviroments, the members of
the clusters should:

* or be accessible directly throught their address,
* or a session affinity header should be configured on the reverse
  proxy.

This is needed since the transaction lives on a single intance of the
cluster and it is not shared between memembers.

By default, the session session affinity header name is
`neo4j-cluster-affinity`.

This header name can be configured in the wrapper creation by using the
`httpSessionAffinityHeader` property on the configuriation.